+++ docs/CHANGELOG.md
@@ -23,6 +23,7 @@
 - added Electric Cleaner control
 - added Punk control, including `O_PUNK_2`, which was unused in OG
 - added Security Guard control
+- added Propeller control
 - changed Ammo Counter to appear in red when the Menu Style is set to PS1
 - changed Punks to have friendliness assignable through Lua, so removing the hard-coded behaviour in the Lud's Gate level sequence
 - changed Trains to no longer hard-code speed based on the level number and instead take it from their default animation